Introduce `executeWithLifecycle` utility for `FlowUseCase` execution
Refactor lifecycle-aware use case operation into shared `executeWithLifecycle` functions for streamlined component integration. Remove duplicate implementations and apply adjustments in affected areas.
