+++ gradle.properties
@@ -1,9 +1,9 @@
 #Gradle
 org.gradle.jvmargs=-Xmx6g -Xms256m -Dfile.encoding=UTF-8 -XX:+UseParallelGC -XX:MaxMetaspaceSize=512m -XX:+HeapDumpOnOutOfMemoryError -Dkotlin.daemon.jvm.options=-XX:MaxMetaspaceSize=1g
 org.gradle.parallel=true
-org.gradle.caching=true
+org.gradle.caching=false
 org.gradle.configureondemand=true
-org.gradle.configuration-cache=true
+org.gradle.configuration-cache=false
 # Use this flag carefully in case some of the plugins are not fully compatible.
 org.gradle.configuration-cache.problems=warn
