+++ shared/arkitekt-decompose/architecture/src/commonMain/kotlin/app/futured/arkitekt/decompose/ext/NavigationExt.kt
@@ -2,7 +2,6 @@ package app.futured.arkitekt.decompose.ext
 
 import com.arkivanov.decompose.router.stack.StackNavigation
 import com.arkivanov.decompose.router.stack.StackNavigator
-import com.arkivanov.decompose.router.stack.bringToFront
 
 /**
  * The same as [StackNavigation.bringToFront] but does not recreate [configuration] if it's class is already on stack and
